Move content of csvslave.py into csvbuilder.py
